requirements_app.txt
gymnasium==1.1.1
Box2D==2.3.10
pygame>=2.6.0
torch>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
streamlit>=1.30.0
Pillow>=10.0.0
pandas>=2.0.0
